+++ src/AssetClass/AssetBase.hs
@@ -100,8 +100,6 @@ data AssetUnion = MO Mortgage
                 deriving (Show, Generic)
 
 
-
-
 $(deriveJSON defaultOptions ''Status)
 $(deriveJSON defaultOptions ''AmortPlan)
 $(deriveJSON defaultOptions ''OriginalInfo)